+++ R/customPSE.R
@@ -179,8 +179,8 @@ customPSE <- function(db, x, xVars, xGrpBy = NULL, xTransform = NULL,
   # different reporting schedules, i.e., if 2016 is most recent in MI and 2017 is
   # most recent in WI, combine them and label as 2017
   if (mr) {
-    xEst <- combineMR(xEst)
-    if (!is.null(y)) yEst <- combineMR(yEst)
+    xEst <- combineMR(xEst, method)
+    if (!is.null(y)) yEst <- combineMR(yEst, method)
   }